(plants): added requireAuth middleware to routes that need authentication

+++ src/routes/plants.js
@@ -12,6 +12,7 @@ import {
   getAvailablePlants,
   getAllPlants,
 } from "../db/plants.js";
+import { requireAuth } from "../middleware/auth.js";
 
 const plantRouter = Router();
 
@@ -34,7 +35,7 @@ plantRouter.get("/all", async (req, res) => {
 });
 
 // GET /plants/:slug
-plantRouter.get("/:slug", async (req, res) => {
+plantRouter.get("/:slug", requireAuth, async (req, res) => {
   const slug = req.params.slug;
 
   const plant = await getPlantBySlug(slug);
@@ -48,9 +49,7 @@ plantRouter.get("/:slug", async (req, res) => {
 })
 
 // POST /plants
-plantRouter.post("/", validatePlant, validatePlantResult, async (req, res) => {
-  // TODO Validation for User and Admin
-  // validatePlant, validatePlantResult,
+plantRouter.post("/", requireAuth, validatePlant, validatePlantResult, async (req, res) => {
 
   const plant = await createPlant(req.body);
 
@@ -58,11 +57,26 @@ plantRouter.post("/", validatePlant, validatePlantResult, async (req, res) => {
 });
 
 // PUT /plants/:slug
-plantRouter.put("/:slug", validatePlant, validatePlantResult, async (req, res) => {
+plantRouter.put("/:slug", 
+  validatePlant, 
+  validatePlantResult,
+  requireAuth,
+  async (req, res) => {
   // TODO Validation for User (owner) and Admin
-
   const slug = req.params.slug;
-  const { name, image, species, lightLevels, coordinates, meetingTime, ownerId } = req.body;
+
+  const plant = await getPlantBySlug(slug);
+
+   if(!plant){
+      return res.status(404).json({ message: "Plant not found" })
+  }
+
+  // Kontrollera att användaren äger plantan
+  if (plant.ownerId._id.toString() !== req.userId) {
+    return res.status(403).json({ message: "Not allowed to update this plant" })
+  }
+
+  const { name, image, species, lightLevels, coordinates, meetingTime } = req.body;
 
   const updatedPlant = await updatePlantBySlug(slug, {
     name,
@@ -86,6 +100,18 @@ plantRouter.put("/:slug", validatePlant, validatePlantResult, async (req, res) =
 plantRouter.patch("/:slug", validatePlantUpdate, validatePlantResult, async (req, res) => {
   // TODO Validation for User (owner) and Admin
   const slug = req.params.slug
+
+  const plant = await getPlantBySlug(slug);
+
+   if(!plant){
+      return res.status(404).json({ message: "Plant not found" })
+  }
+
+  // Kontrollera att användaren äger plantan
+  if (plant.ownerId._id.toString() !== req.userId) {
+    return res.status(403).json({ message: "Not allowed to update this plant" })
+  }
+
   const { ownerId, slug: bodySlug, ...updateData } = req.body
 
   const updatedPlant = await updatePlantBySlug(slug, updateData)
@@ -105,8 +131,18 @@ plantRouter.delete("/:slug", async (req, res) => {
 
   const slug = req.params.slug;
 
+  const findPlant = await getPlantBySlug(slug);
+  
+  if(!findPlant){
+    return res.status(404).json({ message: "Plant not found" })
+  }
+  
+  // Kontrollera att användaren äger plantan
+  if (plant.ownerId._id.toString() !== req.userId) {
+    return res.status(403).json({ message: "Not allowed to update this plant" })
+  }
   const plant = await deletePlantBySlug(slug);
-
+  
   if (!plant) {
     return res.status(400).json({
       message: "Plant does not exist",
